+++ web/src/app/components/table/TableHeader.tsx
@@ -1,5 +1,5 @@
 import { TableHead, TableRow, TableCell } from '@mui/material';
-import { TableSortBy } from '../../../typeDefs/typeDef';
+import { TableSortBy } from '../../../utils/typeDef';
 import { ArrowUpwardOutlined, ArrowDownwardOutlined } from '@mui/icons-material';
 import { useState } from 'react';
 import styled from '@emotion/styled';
